torch&lightning pipelines, predict not yet included
